move JSONSchema to own dir
