+++ packages/quicktype-core/src/language/JSONSchema/index.ts
@@ -1,0 +1,2 @@
+export { JSONSchemaTargetLanguage } from "./language";
+export { JSONSchemaRenderer } from "./JSONSchemaRenderer";
