+++ packages/quicktype-core/src/language/JSONSchema/utils.ts
@@ -1,0 +1,26 @@
+import { funPrefixNamer } from "../../Naming";
+import {
+    legalizeCharacters,
+    splitIntoWords,
+    combineWords,
+    firstUpperWordStyle,
+    allUpperWordStyle
+} from "../../support/Strings";
+
+export const namingFunction = funPrefixNamer("namer", jsonNameStyle);
+
+const legalizeName = legalizeCharacters(cp => cp >= 32 && cp < 128 && cp !== 0x2f /* slash */);
+
+function jsonNameStyle(original: string): string {
+    const words = splitIntoWords(original);
+    return combineWords(
+        words,
+        legalizeName,
+        firstUpperWordStyle,
+        firstUpperWordStyle,
+        allUpperWordStyle,
+        allUpperWordStyle,
+        "",
+        _ => true
+    );
+}
